requirements.txt
asgiref==3.8.1
django==5.1.6
django-debug-toolbar==4.2.0
django-filter==23.5
django-formtools==2.5.1
sqlparse==0.5.3
typing-extensions==4.12.2
pillow==10.3.0
